requirements.txt
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
aiogram>=3.24.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.9
httpx>=0.25.0
